+++ src/luoyi/duke/data/exception/DukeIllegalArgumentException.java
@@ -1,0 +1,10 @@
+package luoyi.duke.data.exception;
+
+/**
+ * A Duke specific exception, thrown when Duke detects illegal/empty argument.
+ */
+public class DukeIllegalArgumentException extends IllegalArgumentException {
+    public DukeIllegalArgumentException(String s) {
+        super(s);
+    }
+}
