--- src/main/java/DukeUnrecognizedArgumentException.java
@@ -1,8 +1,0 @@
-/**
- * A Duke specific exception, thrown when argument cannot be recognized by Duke.
- */
-public class DukeUnrecognizedArgumentException extends IllegalArgumentException {
-    public DukeUnrecognizedArgumentException(String s) {
-        super(s);
-    }
-}
